Remove the WidendBounds map to optimize space [5/n]
Currently, in BoundsAnalysis we save the In sets of all blocks in a map called
WidenedBounds. This is really not needed as we can simply return the In set
for the block which already contains the widened bounds for that block. So we
can get rid of the WidenedBounds map.