Change return type of BuildDeclRefExpr from DeclRefExpr* to ExprResult
Upstream has changed the return type of BuildDeclRefExpr from ExprResult to
DeclRefExpr*. But our Checked C code still makes use of ExprResult. Currently,
I do not have a good way of handling this. So in order to allow us to catch up
to release_90 I am temporarily changing the return type back to ExprResult. We
need to come back and fix this soon. The following ADO task tracks this:
https://dev.azure.com/msresearch/4x4/_workitems/edit/115132