Set the correct offset for ASTStmtReader::NumExprFields
For creating an empty DeclRefExpr, we need to pass NumTemplateArgs. This
depends on the NumExprFields.

In Checked C, we changed the NumExprFields from 5 to 7. But upstream has this
field as 6. So we adjust this and change it to 8.

(cherry picked from commit 9902f10f8a9506823802b86e954e395f25181eb5)