Sort blocks on block IDs instead of using PostOrderCFGView
Currently, we iterate blocks in the topologically sorted order. As a result,
there is no relative ordering between blocks at the same level. For example,
consider the blocks: B1->B3, B2->B3. A topologically sorted order would result
in the order [B1, B3, B2]. This can make the analysis difficult to understand
and debug in case of loops. So we are simplifying the ordering of blocks by
simply sorting them on block IDs which decrease from entry to exit.