Support bounds widening for conditionals with while loops
Added support for widenening bounds for nt_array_ptr dereferences in while
loops. For example, "while (*p) {}" would widen the bounds of p upon
entry to the loop.

In order to handle loops we initialize the In and Out sets of the entry block
as empty. This is done so that we can handle loop back edges and unconditional
jumps.