Fix the handling of killed bounds with loops
The dataflow equation for the Out set is:
  Out[B1->B2] = (In[B1] - Kill[B1]) ∪ Gen[B1->B2]

So if a variable is in the Gen set for an edge then it would be added to the
Out set for the edge even if it is killed inside the outgoing block. This
happens because of the union operation with Gen. This becomes problematic for
loops and we end up widening the bounds in the following case:

  while (p[i])
    p++;

So we handle this as:
  1. Compute the Out set for a block B as usual.
  2. Then, if a variable V is in Kill[B], remove V from the Gen[B->B'].

This would have no effect on liner code but in loops this would ensure we do
not widen bounds in the above case.