Improvements to Checked-C-Convert including Interactive Conversion (#826)
This merges all the changes that people at the Univ. of Maryland and elsewhere have done to the Checked C conversion tool.  Most importantly, instead of doing an automated conversion, it moves in the direction of allowing interactive conversion.  

First, as we want to design in the direction of interactive conversion, we have moved our tool under `clang-tools-extra/clangd/cconvert`. Specifically, we plan to use the language server, i.e., `clangd`, to implement our interactive approach.

Instead of integrating our approach in the design of `clangd`, we use a preprocessor directive `INTERACTIVECCCONV` to build `icconv`, which is a reduced version of `clangd`. 
The preprocessor flag `INTERACTIVECCCONV` separates the code changes such that the original `clangd` is unaffected by our changes and create a language server `icconv`.
The `icconv` is a simple language server that only implements the interactive version of our tool.

We also modified `clang-tools-extra/clangd/cconvert/CMakeLists.txt` to build `CConvertStandalone` which is a standalone version (i.e., a standalone clang tool) of our tool.

The change includes various improvement. Specifically,
* [https://github.com/plum-umd/checkedc-clang/issues/53](https://github.com/plum-umd/checkedc-clang/issues/53)
* [https://github.com/plum-umd/checkedc-clang/issues/52](https://github.com/plum-umd/checkedc-clang/issues/52)
* [https://github.com/plum-umd/checkedc-clang/issues/49](https://github.com/plum-umd/checkedc-clang/issues/49)
* [https://github.com/plum-umd/checkedc-clang/issues/46](https://github.com/plum-umd/checkedc-clang/issues/46)
* [https://github.com/plum-umd/checkedc-clang/issues/45](https://github.com/plum-umd/checkedc-clang/issues/45)
* [https://github.com/plum-umd/checkedc-clang/issues/43](https://github.com/plum-umd/checkedc-clang/issues/43)
* [https://github.com/plum-umd/checkedc-clang/issues/39](https://github.com/plum-umd/checkedc-clang/issues/39)
* [https://github.com/plum-umd/checkedc-clang/issues/33](https://github.com/plum-umd/checkedc-clang/issues/33)
* [https://github.com/plum-umd/checkedc-clang/issues/26](https://github.com/plum-umd/checkedc-clang/issues/26)
* [https://github.com/plum-umd/checkedc-clang/issues/25](https://github.com/plum-umd/checkedc-clang/issues/25)
* [https://github.com/plum-umd/checkedc-clang/issues/24](https://github.com/plum-umd/checkedc-clang/issues/24)
* [https://github.com/plum-umd/checkedc-clang/issues/33](https://github.com/plum-umd/checkedc-clang/issues/33)

Co-authored-by: hasantouma <hasantouma@gmail.com>
Co-authored-by: Aaron Eline <aeline@umd.edu>
Co-authored-by: Aaron Eline <aaron@correctcomputation.com>
Co-authored-by: Ian Sweet <iannsweet@gmail.com>
Co-authored-by: Michael Hicks <mwh@cs.umd.edu>
Co-authored-by: Ian Sweet <ian@correctcomputation.com>