Making checked-c-convert into a library (#837)
This large change adds support for existential struct types to Checked C.  These can be used to replace uses of void pointers in callbacks, abstract data types, and closures.  There is a corresponding commit of tests for existential types to the Checked C repo.

This changes adds IR support, parsing support, semantic checking/typechecking support, and code generation support.  Serialization/deserialization support still needs to be implemented.   There are also a number of cleanups to do, which are documented in GitHub issues.   There are three parts to the change:
- Support for existential types.
- The pack operation.
- The unpack operation, which is done as a declaration with an initializer.

Merged on behalf of Machiry Aravind Kumar (@Machiry)