Semantically compare expressions for equality using a preorder AST
We implement PreoderAST data structure to compare two expressions.
This AST is a mirror of the clang AST with the following properties:
-	It is an n-ary tree
-	It represents an expression in the preorder form
-	Each node of the clang AST is a node in the PreorderAST
-	Its internal nodes represent the operator nodes of the clang AST
-	Its leaf nodes represent the operands of the clang AST

We traverse the ASTs for the expressions E1 and E2 and create the respective
PreorderAST’s. Then we perform tree transforms to "normalize" the expressions
and finally match the tree node-by-node to determine if E1 and E2 are equal.