Support to disable Checked C type checking (#874)
The Checked C conversion tool is built using clang as a library.  We want the Checked C conversion tool to be usable interactively.  In this scenario, the conversion tool will be run and introduce some annotations.  The programmer may then introduce additional annotations and re-run the tool.  This means that the conversion tool needs to support partially annotated code. However, partially annotated code could fail type-checking and consequently, the generated clang AST might not be complete, impeding the tool.  This changes adds a new flag -fcheckedc-convert-tool to the Checked C clang compiler that suppresses diagnostics due to partially annotated Checked C code.  This includes includes type checking errors cause by checked and unchecked pointers mismatches, missing initializers, and missing bounds annotations.

As an example, suppose want to convert the following code:

void foo(_Ptr<int> z) {
  int *x = z;
}

Current clang emits an error:

error: initializing 'int *' with an expression of an incompatible type

      '_Ptr<int>'
  int *x = z;

and the AST produced is incomplete:

./clang -cc1 -ast-dump foo.c:

`-FunctionDecl 0x10f2fbd0 </home/machiry/checkedc/dd.c:1:1, line:3:1> line:1:6 foo 'void (_Ptr<int>)'
  |-ParmVarDecl 0x10f2faf8 <col:10, col:20> col:20 used z '_Ptr<int>'
  `-CompoundStmt 0x10f301c8 <col:23, line:3:1>
    `-DeclStmt 0x10f301b0 <line:2:3, col:13>
      `-VarDecl 0x10f2fd08 <col:3, col:8> col:8 invalid x 'int *'
1 error generated.

When the flag is turned on, diagnostics that might be caused by partially annotated programs are suppressed at the point of diagnostic message generation.  clang doesn't really differentiate whether the cause of an error is syntactic, type checking, or some other semantic checking problem.  It throws away expressions when errors are found, to avoid spurious or confusing error messages. This means that suppressing the errors at the point of generation is the best way to present a complete AST to the conversion tool.

The flag should not be turned on for code generation because the resulting code may be incorrect.   It is only meant to be turned on when ASTs are being created.