Fixes for Visual Studio 16.6.5 update
After updating VS to version 16.6.5, the compiler build on Windows with VS seems to fail. This is because our compiler sources are now a bit outdated and do not have the fixes which the community has for supporting newer versions of VS. This PR brings in two such fixes from the community.

Refer:
1. https://bugs.llvm.org/show_bug.cgi?id=41367
2. https://reviews.llvm.org/D80433
3. https://reviews.llvm.org/D64937