New readmes: Replace two spaces at the end of a sentence with one.
I'm in the minority, and the inconsistency is bound to become
distracting.  Also, there's no good way for an automated rewrapping tool
to know whether a period at the end of a line should get one space or
two when it gets moved.  (Emacs solves this by not breaking after a
period followed by a single space, but the irregular breaking looks
terrible to me.)