Delete more tests that are obsolete (according to Aravind).
This leaves us with the "regression tests" in clang/test/3C (which are
documented in clang/docs/checkedc/3C/CONTRIBUTING.md) as the only tests
of 3C in this source tree.  We have some external tests that we plan to
document when they are in good enough shape to be run by the public.