Compare Expressions: Extract offset from dereference expression [8/n] (#929)
Extract the offset by which a pointer is dereferenced. For the pointer we
compare the dereference expr with the declared upper bound expr. If the
non-integer parts of the two exprs are not equal we say that a valid
offset does not exist and return false. If the non-integer parts of the
two exprs are equal the offset is calculated as:
`(integer part of deref expr - integer part of upper bound expr)`

For example, given upper bound expr `(p + i)` and deref expr `*(p + i + 1)` we
return `1` as the offset.