Compare Expressions: Remove erroneous condition on coalescing nodes
We had erroneously added the condition that to coalesce nodes the binary
operators of the parent and child node must be equal. This fails for the case
where we first constant-fold nodes having a differen operator than the parent
and then need to coalesce the constant folded node into the parent.

Fixes issue #932