Repair mismerge in afffadf85663e6cd26415589cb1b27e4c5ebbb96.
That commit somehow did the equivalent of an ours-merge.  This commit
resets the tree to the result of a merge I did myself.  In that merge,
when I resolved textual conflicts, I attempted to correctly format the
result manually, but there are probably other formatting issues that I
don't want to try to fix at this time.  I'll run full clang-format
passes periodically until we have a better way to automate running them
on every pull request.