Where Clause: Parse a where clause that redeclares bounds [1/n]
Added logic to parse a where clause that redeclares the bounds of a pointer.
For example,

```
_Nt_array_ptr<char> p : count(0);
int x = strlen(p) _Where p : bounds(p, p + x);
```

For the pointer we add a new annotation called WClause in the AST. WClause is
an instance of class WhereClause and holds all facts declared for a pointer
using a where clause.

In this PR:
1. We only parse a where clause declared after a function call, like strlen in
the above example.
2. We only parse facts which redeclare bounds for a checked pointer. For
example, we do not parse facts like `where x == e1`.
3. We only parse a single fact declared using a where clause. For example, we
do not parse multiple facts like `where x > e1 && x < e2`.