Add -base-dir to 3c commands in *rewrite_header tests.
This is the same workaround used in many other tests for a
Windows-specific problem where 3C thinks it isn't allowed to write a
file because (1) it isn't under the base dir (which defaults to
something under the build directory) and (2) the backslash in the
canonicalized file path doesn't match the hard-coded forward slash in
the RUN command.  We should probably fix this silly problem a better
way, but first we may as well apply the workaround we already have.

While I'm here, replace --output-postfix with -output-postfix for
consistency with the other tests.