Make the major _3CInterface methods return false if a tool execution fails.
It's unclear whether (1) there are still more operations in these
methods that can fail and would merit an early return and (2) these
early returns could cause problems for callers such as clangd3c that
want to recover.  But for now, adding these early returns is better than
the status quo, and the only caller we support is the 3c tool, which
just exits when one of these methods returns false.

This exposed a problem with the partial_checked_arr test: it ran 3c on a
lit temporary file named partial_checked_arr.c.tmp, and that file
extension seems to cause 3c to spew errors.  Previously 3c nevertheless
exited 0 and the test carried on; now 3c would exit 1.  Work around this
by using -output-postfix like many of our other tests.  This may not be
the approach we want in the long term, but we don't have the new
approach ready yet.