New more liberal insertion of itypes on internaly unsafe functions (#356)
Implements a new method for inserting itypes that generally results in more total pointers being converted to checked types.
- When a function parameter is used unsafely internal to the function, it is given an itype instead of an wild pointer type
- When a function parameter is used unsafely external to the function, it is given a standard checked pointer type 
- Unsafe arguments passed to otherwise safe functions are cast to the appropriate checked pointer type at call sites

This change is implemented by splitting the internal and external type of function parameters in separate constraint variables that can solve to different checked types. The internal type of a function parameter can solve to wild while the external of type of the pointer is still checked. During rewriting, a function parameter is given an itype if it's internal constraint variable is wild, but it should expose a checked interface due to a checked external constraint variable.