Disable assertions checking for constraint validity (#379)
Disable checks that were used to verify that the relationships between constant
atom were valid. While we do not generally intend to rewrite code in a way that
violates these relationships, it is possible for a user of 3C to manually
modify their code so that it does while still having valid CheckedC code.