Fix bug in PR #380
CheckedC does let you take the address of constant size checked arrays. The
previous commit would not let this happen, causing the ptrtoconstarr.c
test to fail. It also caused ptrptr.c to fail, but this was because the
-alltypes behavior expected in that test generated code that did not
compile (because it took the address of an _Array_ptr). Both tests are
now passing.