Handle function declaration appearing after definition (fix #399) (#411)
Ensure that constrains generated due to a function's declaration are
applied to the variables generated at the definition even when
declaration appears after the definition. This is done by copying
the atoms from the definition constraint variable into the declaration
constraint variable using brainTransplant before any constraints are
added on the declaration.

Also change insertNewFVConstraint to return void. Previously, it
returned a bool, but the value was never used. This also
applies to two functions that are only ever called from inside
insertNewFVConstraint.