Remove special handling of functions defined in system headers. (#431)
The liberal itypes change (#356) added code to flag them as not having a
body even though they did. This may have been a useful mitigation for
certain problems with unwritable files at the time, but it risks causing
confusion and violating 3C invariants in the future. From now on, we
want to handle all unwritable files (including system headers) via the
new canWrite constraints approach.

It appears that removing this code will immediately fix one problem with
cast insertion that is currently blocking the benchmark tests (#423), so
we'll go ahead and make this change even though we don't fully
understand the cast insertion problem or how to appropriately
regression-test it.