More consistent handling of rewriting around macros (fix #429) (#434)
Previously a macro appearing at the beginning of rewritten source range
would not cause failure in rewriting, but a macro coming at the end of
the range would. This was because the "expansion" source location was
used for the start of the range, but it was not used for the end.

In some parts of the rewriter, the expansion location was not used for
either end of the range. These have been changed to use the same logic
as the other rewritings. This change will allow some rewritings that
would have been silently dropped and emit an error on others where no
rewriting is possible.

The rewriting function now uses makeFileCharRange instead of the function
getExpansionRange. This lets 3C emit an error where the macro expansion 
of the range included tokens that are not part of what we want to replace.
Previously 3C would silently delete these tokens.

Co-authored-by: Matt McCutchen (Correct Computation) <matt@correctcomputation.com>