Fix incorrect rewritting of function with typdefed type (fix #430) (#436)
Fixes a error in rewriting for function declaration such as

    typedef void foo(int*);
    foo bar;
    void bar(int *a){};

Previously `FunctionDeclBuilder` only replaced the parameter declaration
of `bar` because nothing in the return type needed to be rewritten. This
caused the rewriting of the first declaration to be missing the return
type. With this change, `FunctionDeclBuilder` detects that `bar` is
declared using a typedef type, and generates replacement text for the
entire declaration instead of only the parameter list. This change does not
implement proper constraint generation for these function, so the typedef foo
is not rewritten. 

Reconstruct a parameter declaration if we can't get its original source
due to a macro.

Co-authored-by: Matt McCutchen (Correct Computation) <matt@correctcomputation.com>