Guard only an assert with NDEBUG check (#990)
* Guard only an assert with NDEBUG check

In a release without asserts build, we hit the error "function type mismatch".
This was because of the code to determine the type of a _For_any function from
the CallInfo was being protected by the #ifndef NDEBUG flag and hence did not
execute. The main purpose of that check is to guard the assert at the end of
that code. So we moved the check to just before the assert.

* Added note about merge conflicts