Fix placement of _Assume_bounds_cast on c-style cast (#447)
The if statement that tries to replace an existing cast with a new cast
assumed that the new cast was a c-style cast and not a CheckedC
bounds cast. If the branch was taken when inserting a bounds cast,
the leading '_' was cut off, resulting in invalid code. This commit
adds a condition that ensures this does not happen.