[Where Clause] Parse a Checked C where clause on a null statement [1/n] (#960)
Parse a Checked C where clause attached to a null statement. Such a where clause would have the following syntax:
_Where fact1 _And fact2

A fact could either be a bounds fact or a relational operator fact. Facts occurring in a where clause are separated by a new keyword _And.

As a follow-up to this PR I will add support for where clauses occurring with declarations and assignments.