Avoid trying to insert cast inside macros (#491)
The constraint created between a function parameter and a corresponding
argument is updated to be an equality constraint when the function call
is inside a macro. Doing this forces the checked type of the parameter
to the same as that of the argument. Casts are only inserted when
parameter and argument types are not equal, so this changes ensures that
3C will not try (and subsequently fail) to insert casts on expressions
in macros.

After the previous change there were still issues with cast insertion due
to source location collision in the ExprConstraintVars map. To avoid
these problems the map has been updated to use the AST node ID instead
of the source location. AST node IDs are only unique within a
translation unit, so they are paired with the name of the main file for
the translation unit.

AST nodes IDs are consistent between multiple clang tool invocations,but
they can change when run on different computers.
http://clang-developers.42468.n3.nabble.com/Question-about-hashing-AST-nodes-td4063810.html#a4063828