Fix function pointer rewriting with itypes (fix #498) (#506)
The code for deciding if a function should get an itype was duplicated
for function declarations and function pointer types. The function
pointer version of the code had a bug in it that caused issue #498.
The duplicated code has been extracted into a pair of functions that are
reused for functions and function pointers.

A lot of the lines changed in this PR are caused by an EnvironmentMap&
parameter being changed to Constraints&. This lets the function pointer
code call solutionEqualTo which is needed for correct itype insertion.