Make `3c -verify` an error for now. (#519)
PR #488 made `3c -verify` cover only the compiler diagnostics, but none
of our regression tests actually use that functionality. Instead, one
regression test (macro_function_call) used `3c -verify` to try to test
the absence of 3C warnings, and we were unaware that the test wasn't
testing what it was supposed to. I think it's best to make `3c -verify`
an error for now so we don't make that mistake again.