Update types for pointers with itypes  (#539)
With this change, parameter variables declared with an itype are treated the
same as unchecked pointers while solving for output pointer types. This is done
by allocating VarAtoms for the ConstraintVariable instead of ConstAtoms. 

A change to how we decide if the type of a pointer has changed was required to
have this work properly. We previously assumed that a pointer had changed (and
therefore needed to be rewritten) if its type in the source code was unchecked,
but the type in the output needed to be checked. That doesn't work anymore
because an itype might change from one pointer type to another. I've added a
new field into the PVConstraint class called SrcVars that holds the ConstAtoms
representing the pointers type in the source code. After constraint solving,
the variable has a new vector of ConstAtoms representing the output pointer
type. If these vectors are different, then the type has changed.

While adding this new field, I took the opportunity to move some code from
ConstraintResolver.cpp into ConstraintVariables.cpp. This moderately improves
encapsulation of the PVConstraint class.

A number of other code changes were required to handle cases where an itype
should not be re-solved. For example, functions without a definition or whose
body is in unwritable code. For these functions we should continue treating
them as they are treated before this PR.