[Where Clause] Dataflow analysis for bounds widening with where clauses [5/n]
This PR adds the interface for the dataflow analysis for bounds widening of
null-terminated arrays, in presence of where clauses. We declare types, classes
and functions used by the analysis.

We declare the BoundsWideningAnalysis class that represents the dataflow
analysis for bounds widening. The sets In, Out, Gen and Kill that are used by
the analysis are members of this class. The class also has methods that act on
these sets to perform the dataflow analysis.