Remove clangd3c.
It is unmaintained, and its implementation in the form of changes to
existing clangd files causes hassles during LLVM baseline upgrades.

The addition of clangd3c was the only difference in the
clang-tools-extra/clangd directory between the baseline and the Checked
C repository, so this change leaves that directory identical to the
baseline.

Also remove the mention of clangd3c from the 3C documentation, but not
other 3C utility code related to clangd3c (that will be addressed
later).