Fix intermittent error bounds inference (fix #585) (#589)
Iterate over a BoundsKey set instead of a ProgramVar* set so that the
iteration order does not depend on the address of the program variables.
This fixes the issue, but it is still potentially a problem that the
function depends on the order it iterations over the bounds keys. It is
also likely that other loops over sets of pointers exbibit this same
problem.