[BoundsVars] Store separate lower and upper bounds variables (#1056)
* [BoundsVars] Store separate lower and upper bounds variables

We separate out BoundsVars into BoundsVarsLower and BoundsVarsUpper that track
the variables occurring in the lower and upper bounds expressions,
respectively. This is needed for the bounds widenening analysis where we want
to determine the variables that can potentially be widened in a given
dereference expression.

* [BoundsWidening] Rename BoundsMapTy to BoundsOffsetMapTy

We rename BoundsMapTy to BoundsOffsetMapTy as this typename will conflict with
the definition in BoundsWideningAnalysis.h.
Note: BoundsWideningAnalysis(.cpp,.h) is meant to replace
BoundsAnalysis(.cpp,.h).