Fix regression tests to not create temp files outside the lit temp directory. (#624)
* Follow new guideline: Avoid `cd DIR && FOO` in RUN lines.

Instead, put `FOO` on a separate run line. `cd` affects later RUN lines,
and `cd DIR && FOO` could mislead the reader that it doesn't and that's
why we needed to put both parts on one line. I added the proposed
guideline to #346.

I'm changing the existing tests before changing more tests to use `cd`
as part of this PR.

* Add `-o /dev/null` to clang commands that use `-` as input.

* Add -working-directory to clang commands with named input files.

* `cd` to a temporary directory for `3c -dump-stats` commands.