Add a flag to always add itypes on function prototypes (#637)
This is the implementation of the -itypes-for-externs flag used in the vsftpd
tutorial. The flag enables a new mode where 3C will emit itypes for all checked
external functions. This should let the headers files containing prototypes for
these function be used indecently from the converted function bodies, which is
intended to enable conversion of individual files in a program without
committing to partial conversion of every file.

This change also adds code to support emitting itypes of structure fields. For
now, this is only done when the flag is supplied, but we might like to support
solving these itypes in the future.

Similarly, it adds support for itypes on parameter's with typedef'ed types. The
typedef is left unchecked, but they are given a checked itype. Again, this is
only done when the new flag is passed, but we might consider using itypes like
this in normal operation of 3C.