[PreorderAST] Add method to compute integer diff between two expressions
Added a method called PreorderAST::GetExprIntDiff that computes the integer
difference between two input expressions and returns true if the expressions
are comparable. If the expressions are not comparable then it returns false.

This method will be used in the updated bounds widening analysis. This method
is meant to replace PreorderAST::GetDerefOffset after the updated bounds
widening analysis merges.

See #1078