If the input argument Stmt *S in the Check function in SemaBounds.cpp is not of any StmtClass but is of type Expr*, then the function Check checks if it is a NullPointerConstant. The function isNullPointerConstant returns true or false depending on what the expression is, but it expects the expression to not be value-dependent.
Now, in the case of the testcase Sema/warn-unused-variables.c, the function isNullPointerConstant
is called on a BlockExpr* which is value-dependent because it has an erroneous Return statement
nested within a compound statement. (The erroneous Return statement contains a subexpression
that is a RecoveryExpr).

The fact that the BlockExpr is value-dependent causes the compiler to reach an llvm_unreachable
state in the function isNullPointerConstant. Therefore the fix is to call the function
isNullPointerConstant only on expressions that are not value-dependent.

In LLVM/Clang 12.x, an erroneous subexpression (i.e. a RecoveryExpr) causes the compiler to create
a value-dependent BlockExpr whereas in LLVM/Clang 11.x, it caused the compiler to create a
non-value-dependent BlockExpr that did not contain the erroneous Return statement.